chore: [ISSUE-010] Auth 패키지를 Member 패키지 하나로 단일화
