+++ src/main/java/com/zero/bwtableback/member/service/MemberService.java
@@ -1,8 +1,4 @@
 package com.zero.bwtableback.member.service;
 
-import com.zero.bwtableback.member.dto.SignupFormDto;
-import com.zero.bwtableback.member.entity.Member;
-
 public interface MemberService {
-    Member signupMember(SignupFormDto form);
 }
